tested move logic, cleaned up piece bit board access per player, updated run command, began implementing pieces, more utility functions and enums for easier readability / testing

+++ Pieces/Piece/piece.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 using std::ostream;
 
-Piece::Piece(int8_t type, int8_t color, int16_t position) {
+Piece::Piece(int8_t type, int8_t color, uint64_t position) {
     this->type = type;
     this->color = color;
     this->position = position;
@@ -17,12 +17,12 @@ int8_t Piece::getColor() {
     return color;
 }
 
-int16_t Piece::getPosition() {
+uint64_t Piece::getPosition() {
     return position;
 }
 
-vector<int16_t> Piece::getMoves() {
-    return vector<int16_t>();
+vector<int32_t> Piece::getMoves() {
+    return vector<int32_t>();
 }
 
 std::ostream& operator<<(std::ostream& os, const Piece& piece) {
