+++ Pieces/Piece/piece.h
@@ -5,7 +5,7 @@
 #include <vector>
 using std::int8_t;
 using std::int16_t;
-using std::int64_t;
+using std::uint64_t;
 using std::string;
 using std::vector;
 
@@ -13,13 +13,13 @@ class Piece {
     public:
     int8_t type;
     int8_t color;
-    int16_t position;
+    uint64_t position;
 
-    Piece(int8_t type, int8_t color, int16_t position);
+    Piece(int8_t type, int8_t color, uint64_t position);
     int8_t getType();
     int8_t getColor();
-    int16_t getPosition();
-    virtual vector<int16_t> getMoves();
+    uint64_t getPosition();
+    virtual vector<int32_t> getMoves();
     friend std::ostream& operator<<(std::ostream& os, const Piece& piece);
 };
